# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ДомДело API
  version: 0.1.0
  description: API MVP для коллективных дел жителей дома
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session
    demoUser:
      type: apiKey
      in: header
      name: X-Demo-User
...
paths:
...
  /api/auth/max:
    post:
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - initData
              properties:
                initData:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Default Response
...
  /api/files/{key}:
    get:
      parameters:
        - schema:
            type: string
          in: path
          name: key
          required: true
      responses:
        "200":
          description: Default Response
servers:
  - url: http://localhost:8080
...